reorganize mask_rcnn training code more into different files
